move tests out of test_validators.py
